Chatbot/requirements.txt
streamlit==1.28.1
google-generativeai==0.3.0
python-dotenv==1.0.0
